+++ packages/markdownit-plugins/jest.config.js
@@ -1,0 +1,24 @@
+module.exports = {
+  "roots": [
+    "<rootDir>/src",
+  ],
+  "transform": {
+    "^.+\\.tsx?$": "ts-jest"
+  },
+  "testURL": "http://localhost/someapphosted?appName=my-mock-app",
+  "testRegex": "(/__tests__/.*|(\\.|/)(test|spec))\\.tsx?$",
+  "moduleFileExtensions": [
+    "ts",
+    "tsx",
+    "js",
+    "jsx",
+    "json",
+    "node"
+  ],
+  "coverageReporters": [
+    "json",
+    "lcov",
+    "text",
+    "cobertura"
+  ],
+}