+++ packages/markdownit-plugins/src/main.ts
@@ -1,0 +1,14 @@
+import {
+  pluginLTLinkTarget,
+  pluginLTPublicMethod,
+  pluginLTAppsSDKMethod,
+  pluginLTAnalytics,
+} from './plugins';
+
+export default function applyLetsTalkPlugins(tokens: ObjectIndex<Token>, idx: string): ObjectIndex<Token> {
+  let newTokens = pluginLTLinkTarget(tokens, idx, '_blank');
+  newTokens = pluginLTPublicMethod(newTokens, idx);
+  newTokens = pluginLTAnalytics(newTokens, idx);
+  newTokens = pluginLTAppsSDKMethod(newTokens, idx);
+  return newTokens;
+}