+++ packages/widgets-manager/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased [2.3.0] - 2020-01-23
+### Added
+- Added method `getMountedApps` to the manager
+
 ## Hotfix [2.2.1, 2.2.2, 2.2.3, 2.2.4] - 2020-01-05
 ### Fixed
 - Disable the Sync of firebase with redux -> Make a TODO to figure out corner